Add 64-bit registers and instructions to MIPS

+++ simulator/mips/mips.h
@@ -2,7 +2,7 @@
  * mips.h - all the aliases to MIPS ISA
  * @author Aleksandr Misevich
  * Copyright 2018 MIPT-MIPS
- */ 
+ */
 
 #ifndef MIPS_H_
 #define MIPS_H_
@@ -15,8 +15,8 @@ struct MIPS
     using FuncInstr = MIPSInstr;
     using Register = MIPSRegister;
     using Memory = InstrMemory<MIPSInstr>;
-    using RegisterUInt = uint32;
-    using RegDstUInt = doubled_t<uint32>; // MIPS may produce output to 2x HI/LO register
+    using RegisterUInt = uint64;
+    using RegDstUInt = doubled_t<uint64>; // MIPS may produce output to 2x HI/LO register
 };
 
 #endif // MIPS_H_
